+++ chain-core/src/property.rs
@@ -30,6 +30,7 @@
 //! is selected to write a block in the chain.
 //!
 
+pub use chain_ser::deser::*;
 use std::{fmt::Debug, hash::Hash};
 
 /// Trait identifying the block identifier type.
@@ -285,27 +286,6 @@ pub trait Settings {
     fn block_version(&self) -> <Self::Block as Block>::Version;
 }
 
-/// Define that an object can be written to a `Write` object.
-pub trait Serialize {
-    type Error: std::error::Error + From<std::io::Error>;
-
-    fn serialize<W: std::io::Write>(&self, writer: W) -> Result<(), Self::Error>;
-
-    /// Convenience method to serialize into a byte vector.
-    fn serialize_as_vec(&self) -> Result<Vec<u8>, Self::Error> {
-        let mut data = vec![];
-        self.serialize(&mut data)?;
-        Ok(data)
-    }
-}
-
-/// Define that an object can be read from a `Read` object.
-pub trait Deserialize: Sized {
-    type Error: std::error::Error + From<std::io::Error> + Send + Sync + 'static;
-
-    fn deserialize<R: std::io::BufRead>(reader: R) -> Result<Self, Self::Error>;
-}
-
 /// Defines the way to parse the object from a UTF-8 string.
 ///
 /// This is like the standard `FromStr` trait, except that it imposes
@@ -326,14 +306,6 @@ where
 
     fn from_str(s: &str) -> Result<Self, Self::Error> {
         std::str::FromStr::from_str(s)
-    }
-}
-
-impl<T: Serialize> Serialize for &T {
-    type Error = T::Error;
-
-    fn serialize<W: std::io::Write>(&self, writer: W) -> Result<(), T::Error> {
-        (**self).serialize(writer)
     }
 }
