+++ chain-ser/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "chain-ser"
+version = "0.1.0"
+authors = ["Vincent Hanquez <vincent.hanquez@iohk.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
